requirements.txt
langgraph==0.6.8
langchain==0.3.27
langchain-openai>=0.3.0
langchain-community==0.3.30
python-dotenv==1.1.1
